+++ pcsx2/x86/microVU_Execute.inl
@@ -354,7 +354,7 @@ _mVUt void mVUcleanUp()
 		u32 cycles_passed = std::min(mVU.cycles, 3000) * EmuConfig.Speedhacks.EECycleSkip;
 		if (cycles_passed > 0)
 		{
-			s32 vu0_offset = VU0.cycle - cpuRegs.cycle;
+			s64 vu0_offset = VU0.cycle - cpuRegs.cycle;
 			cpuRegs.cycle += cycles_passed;
 
 			// VU0 needs to stay in sync with the CPU otherwise things get messy
